ENH: Add the remote as a parameter for indexing a given server
Add the remote as a parameter for indexing a given server.

The remote is then stored along with the rest of the data so that the
subsequent subcommands know what server they need to query. For now, we
rely on the known OpenNeuro attributes.

Incidentally, fix a bug that made such that the cursors that did not
contain any content were being counted as valid indexed datasets and
and serialized to the output file containing the `None` string. Filter
those cursors.