FIX: Use the GraphQL `key` field to query directories
Use the GraphQL `key` field to query directories: the contents of the
directories were no longer being listed when collecting (`niquery
collect`) the dataset information, and only the files at the root of the
datasets (e.g. `CHANGES`, `dataset_description.json`, etc.) were being
listed. This made such that the no NIfTI files were being listed for
subset selection purposes.

This patch set employs the `key` field of the returned GraphQL response
to recursively query the directories. This allows them to be effectively
traversed, and the NIfTI files to be listed and written in the dataset
content TSV files.