ENH: Read multi-line columns when filtering modality records
Read multi-line columns when filtering modality records: set the data
type that is expected when reading the TSV file that contains the list
of files and their attributes for a given dataset.

It turns out that the `urls` column stored in those files may contain
several values all pointing to the same file. Although the field is
slated to be a list, it contains a single value for virtually all
retrieved records from OpenNeuro. There is, however, a dataset
(`ds006833`) for which its record with `id`
`bd4762f8223a0dfd69627dfff148c8c50020ccdc` and `filename`
`sub-01_ses-02_task-DeepMReyeCalib_run-01_recording-eye1_physioevents.json`
has a `urls` value that contains 5 items. Trying to analyze the dataset:
```
analyze \
/path/to/dataset_files \
/path/to/dataset_features \
--suffix bold
```

causes a segmentation fault when trying to read the file:
```
Segmentation fault (core dumped)
```

or
```
Process finished with exit code 139 (interrupted by signal 11:SIGSEGV)
```

when being debugged through the IDE.

The underlying error triggered by `pandas` reads:
```
pandas.errors.ParserError
Error tokenizing data. C error: Expected 1 fields in line 41, saw 5')
```

So the issue is caused by a malformed content from the point of view of
`pandas`.

This patch set circumvents the problem by forcing to read the data type
of the column at issue (`urls`) as a string (`str`) using the `pandas`
`converter` parameter. The data type for the preceding column needs to
be specified as well, so this patch set specifies the data types for
all columns.

Define all necessary column names in the `utils.attributes.py` module.